renaming chart file names
