+++ app/javascript/components/charts/cp-performance-all-sectors/Chart.js
