+++ app/javascript/components/charts/cp-performance/Chart.js
